REFACTOR : 폴더/파일 구조를 FSD구조로 변경
