+++ src/features/login/model/type.ts
@@ -1,0 +1,4 @@
+export interface LoginType {
+    email: string;
+    password: string;
+}
