+++ src/features/login/ui/LoginForm.module.css
@@ -1,0 +1,19 @@
+.form {
+    @apply w-full flex flex-col gap-3 p-4;
+}
+
+.login {
+    @apply w-full;
+}
+
+.google {
+    @apply w-full;
+}
+
+.signup {
+    @apply w-full bg-transparent text-black;
+
+    &:hover {
+        @apply bg-slate-200;
+    }
+}
