+++ src/features/signup/api/api.tsx
@@ -1,0 +1,22 @@
+import { SignupType } from '@/features/signup/model/type';
+import supabase from '@/supabaseClient';
+
+export const onSignup = async (form: SignupType) => {
+    // auth
+    const { data, error } = await supabase.auth.signUp({
+        ...form,
+    });
+
+    if (error) {
+        throw error;
+    }
+
+    await supabase.from('users').insert({
+        id: data.user?.id, // 회원가입 성공 시 받아온 data중 id(uid) 값을 가져온다.
+        email: data.user?.email,
+        created_at: data.user?.created_at,
+        nickname: form.nickname,
+        password: form.password,
+        isSeller: form.isSeller,
+    });
+};
