+++ src/features/signup/ui/SignupForm.module.css
@@ -1,24 +1,3 @@
-.header {
-    @apply p-3 flex gap-2;
-    background-color: rgb(242, 242, 242);
-    font-weight: bold;
-
-    svg {
-        cursor: pointer;
-    }
-}
-
-.layout {
-    @apply w-full h-full flex flex-col gap-2;
-    margin: 0 auto;
-    border-right: 1px solid #ccc;
-    border-left: 1px solid #ccc;
-
-    @media all and (min-width: 1024px) {
-        width: 600px;
-    }
-}
-
 .form {
     @apply w-full flex flex-col gap-3 p-6;
 }
