+++ src/pages/HomePage/ui/HomePage.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/ui/button';
+import { Button } from '@/shared/components/ui/button';
 import supabase from '@/supabaseClient';
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
