+++ src/pages/LoginPage/ui/LoginPage.module.css
@@ -15,23 +15,3 @@
     border-right: 1px solid #ccc;
     border-left: 1px solid #ccc;
 }
-
-.form {
-    @apply w-full flex flex-col gap-3 p-4;
-}
-
-.login {
-    @apply w-full;
-}
-
-.google {
-    @apply w-full;
-}
-
-.signup {
-    @apply w-full bg-transparent text-black;
-
-    &:hover {
-        @apply bg-slate-200;
-    }
-}
