+++ src/pages/SignupPage/index.ts
@@ -1,0 +1,1 @@
+export { SignupPage } from './ui/SignupPage';
