+++ src/pages/SignupPage/ui/SignupPage.module.css
@@ -1,0 +1,20 @@
+.header {
+    @apply p-3 flex gap-2;
+    background-color: rgb(242, 242, 242);
+    font-weight: bold;
+
+    svg {
+        cursor: pointer;
+    }
+}
+
+.layout {
+    @apply w-full h-full flex flex-col gap-2;
+    margin: 0 auto;
+    border-right: 1px solid #ccc;
+    border-left: 1px solid #ccc;
+
+    @media all and (min-width: 1024px) {
+        width: 600px;
+    }
+}
